requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
numpy>=1.19.5
pandas>=1.1.5
scikit-learn>=0.24.2
textblob>=0.15.3
nltk>=3.6.2
pytesseract>=0.3.8
pillow>=8.2.0
opencv-python>=4.5.1
PyPDF2>=2.0.0
python-docx>=1.1.2
openpyxl>=3.1.5
xlrd>=2.0.1
waitress>=3.0.2
